feat(admin): ajout de la page d'inscription en masse avec gestion des fichiers, des erreurs et des mappings
